+++ jms/src/main/scala/akka/stream/alpakka/jms/impl/JmsConnector.scala
@@ -4,7 +4,6 @@
 
 package akka.stream.alpakka.jms.impl
 
-import java.util.concurrent.ArrayBlockingQueue
 import java.util.concurrent.atomic.AtomicReference
 
 import akka.actor.ActorSystem
@@ -320,238 +319,5 @@ private[jms] object JmsConnector {
     case JmsConnectorInitializing(c, _, _, _) => c
     case JmsConnectorConnected(c) => Future.successful(c)
     case _ => Future.failed(JmsNotConnected)
-  }
-}
-
-/**
- * Internal API.
- */
-@InternalApi
-private[jms] trait JmsConsumerConnector extends JmsConnector[JmsConsumerSession] {
-  this: TimerGraphStageLogic with StageLogging =>
-
-  override val startConnection = true
-
-  protected def createSession(connection: jms.Connection,
-                              createDestination: jms.Session => jms.Destination): JmsConsumerSession
-
-}
-
-/**
- * Internal API.
- */
-@InternalApi
-private[jms] trait JmsProducerConnector extends JmsConnector[JmsProducerSession] {
-  this: TimerGraphStageLogic with StageLogging =>
-
-  protected final def createSession(connection: jms.Connection,
-                                    createDestination: jms.Session => jms.Destination): JmsProducerSession = {
-    val session = connection.createSession(false, AcknowledgeMode.AutoAcknowledge.mode)
-    new JmsProducerSession(connection, session, createDestination(session))
-  }
-
-  override val startConnection = false
-
-  val status: JmsProducerMatValue = new JmsProducerMatValue {
-    override def connected: Source[JmsConnectorState, NotUsed] =
-      Source.fromFuture(connectionStateSource).flatMapConcat(identity)
-  }
-}
-
-/**
- * Internal API.
- */
-@InternalApi
-private[jms] object JmsMessageProducer {
-  def apply(jmsSession: JmsProducerSession, settings: JmsProducerSettings, epoch: Int): JmsMessageProducer = {
-    val producer = jmsSession.session.createProducer(null)
-    if (settings.timeToLive.nonEmpty) {
-      producer.setTimeToLive(settings.timeToLive.get.toMillis)
-    }
-    new JmsMessageProducer(producer, jmsSession, epoch)
-  }
-}
-
-/**
- * Internal API.
- */
-@InternalApi
-private[jms] class JmsMessageProducer(jmsProducer: jms.MessageProducer,
-                                      jmsSession: JmsProducerSession,
-                                      val epoch: Int) {
-
-  private val defaultDestination = jmsSession.jmsDestination
-
-  private val destinationCache = new SoftReferenceCache[Destination, jms.Destination]()
-
-  def send(elem: JmsMessage): Unit = {
-    val message: jms.Message = createMessage(elem)
-    populateMessageProperties(message, elem)
-
-    val (sendHeaders, headersBeforeSend: Set[JmsHeader]) = elem.headers.partition(_.usedDuringSend)
-    populateMessageHeader(message, headersBeforeSend)
-
-    val deliveryMode = sendHeaders
-      .collectFirst { case x: JmsDeliveryMode => x.deliveryMode }
-      .getOrElse(jmsProducer.getDeliveryMode)
-
-    val priority = sendHeaders
-      .collectFirst { case x: JmsPriority => x.priority }
-      .getOrElse(jmsProducer.getPriority)
-
-    val timeToLive = sendHeaders
-      .collectFirst { case x: JmsTimeToLive => x.timeInMillis }
-      .getOrElse(jmsProducer.getTimeToLive)
-
-    val destination = elem.destination match {
-      case Some(messageDestination) => lookup(messageDestination)
-      case None => defaultDestination
-    }
-    jmsProducer.send(destination, message, deliveryMode, priority, timeToLive)
-  }
-
-  private def lookup(dest: Destination) = destinationCache.lookup(dest, dest.create(jmsSession.session))
-
-  private[jms] def createMessage(element: JmsMessage): jms.Message =
-    element match {
-
-      case textMessage: JmsTextMessage => jmsSession.session.createTextMessage(textMessage.body)
-
-      case byteMessage: JmsByteMessage =>
-        val newMessage = jmsSession.session.createBytesMessage()
-        newMessage.writeBytes(byteMessage.bytes)
-        newMessage
-
-      case mapMessage: JmsMapMessage =>
-        val newMessage = jmsSession.session.createMapMessage()
-        populateMapMessage(newMessage, mapMessage)
-        newMessage
-
-      case objectMessage: JmsObjectMessage => jmsSession.session.createObjectMessage(objectMessage.serializable)
-
-    }
-
-  private[jms] def populateMessageProperties(message: javax.jms.Message, jmsMessage: JmsMessage): Unit =
-    jmsMessage.properties.foreach {
-      case (key, v) =>
-        v match {
-          case v: String => message.setStringProperty(key, v)
-          case v: Int => message.setIntProperty(key, v)
-          case v: Boolean => message.setBooleanProperty(key, v)
-          case v: Byte => message.setByteProperty(key, v)
-          case v: Short => message.setShortProperty(key, v)
-          case v: Long => message.setLongProperty(key, v)
-          case v: Double => message.setDoubleProperty(key, v)
-          case null => throw NullMessageProperty(key, jmsMessage)
-          case _ => throw UnsupportedMessagePropertyType(key, v, jmsMessage)
-        }
-    }
-
-  private def populateMapMessage(message: javax.jms.MapMessage, jmsMessage: JmsMapMessage): Unit =
-    jmsMessage.body.foreach {
-      case (key, v) =>
-        v match {
-          case v: String => message.setString(key, v)
-          case v: Int => message.setInt(key, v)
-          case v: Boolean => message.setBoolean(key, v)
-          case v: Byte => message.setByte(key, v)
-          case v: Short => message.setShort(key, v)
-          case v: Long => message.setLong(key, v)
-          case v: Double => message.setDouble(key, v)
-          case v: Array[Byte] => message.setBytes(key, v)
-          case null => throw NullMapMessageEntry(key, jmsMessage)
-          case _ => throw UnsupportedMapMessageEntryType(key, v, jmsMessage)
-        }
-    }
-
-  private def populateMessageHeader(message: javax.jms.Message, headers: Set[JmsHeader]): Unit =
-    headers.foreach {
-      case JmsType(jmsType) => message.setJMSType(jmsType)
-      case JmsReplyTo(destination) => message.setJMSReplyTo(destination.create(jmsSession.session))
-      case JmsCorrelationId(jmsCorrelationId) => message.setJMSCorrelationID(jmsCorrelationId)
-    }
-}
-
-/**
- * Internal API.
- */
-@InternalApi
-private[jms] sealed trait JmsSession {
-
-  def connection: jms.Connection
-
-  def session: jms.Session
-
-  private[jms] def closeSessionAsync()(implicit ec: ExecutionContext): Future[Unit] = Future { closeSession() }
-
-  private[jms] def closeSession(): Unit = session.close()
-
-  private[jms] def abortSessionAsync()(implicit ec: ExecutionContext): Future[Unit] = Future { abortSession() }
-
-  private[jms] def abortSession(): Unit = closeSession()
-}
-
-/**
- * Internal API.
- */
-@InternalApi
-private[jms] class JmsProducerSession(val connection: jms.Connection,
-                                      val session: jms.Session,
-                                      val jmsDestination: jms.Destination)
-    extends JmsSession
-
-/**
- * Internal API.
- */
-@InternalApi
-private[jms] class JmsConsumerSession(val connection: jms.Connection,
-                                      val session: jms.Session,
-                                      val jmsDestination: jms.Destination,
-                                      val settingsDestination: Destination)
-    extends JmsSession {
-
-  private[jms] def createConsumer(
-      selector: Option[String]
-  )(implicit ec: ExecutionContext): Future[jms.MessageConsumer] =
-    Future {
-      (selector, settingsDestination) match {
-        case (None, t: DurableTopic) =>
-          session.createDurableSubscriber(jmsDestination.asInstanceOf[jms.Topic], t.subscriberName)
-
-        case (Some(expr), t: DurableTopic) =>
-          session.createDurableSubscriber(jmsDestination.asInstanceOf[jms.Topic], t.subscriberName, expr, false)
-
-        case (Some(expr), _) =>
-          session.createConsumer(jmsDestination, expr)
-
-        case (None, _) =>
-          session.createConsumer(jmsDestination)
-      }
-    }
-}
-
-/**
- * Internal API.
- */
-@InternalApi
-private[jms] class JmsAckSession(override val connection: jms.Connection,
-                                 override val session: jms.Session,
-                                 override val jmsDestination: jms.Destination,
-                                 override val settingsDestination: Destination,
-                                 val maxPendingAcks: Int)
-    extends JmsConsumerSession(connection, session, jmsDestination, settingsDestination) {
-
-  private[jms] var pendingAck = 0
-  private[jms] val ackQueue = new ArrayBlockingQueue[() => Unit](maxPendingAcks + 1)
-
-  def ack(message: jms.Message): Unit = ackQueue.put(message.acknowledge _)
-
-  override def closeSession(): Unit = stopMessageListenerAndCloseSession()
-
-  override def abortSession(): Unit = stopMessageListenerAndCloseSession()
-
-  private def stopMessageListenerAndCloseSession(): Unit = {
-    ackQueue.put(() => throw StopMessageListenerException())
-    session.close()
   }
 }
