+++ jms/src/main/scala/akka/stream/alpakka/jms/impl/JmsMessageProducer.scala
@@ -1,0 +1,123 @@
+/*
+ * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
+ */
+
+package akka.stream.alpakka.jms.impl
+
+import akka.annotation.InternalApi
+import akka.stream.alpakka.jms._
+import javax.jms
+
+/**
+ * Internal API.
+ */
+@InternalApi
+private[jms] class JmsMessageProducer(jmsProducer: jms.MessageProducer,
+                                      jmsSession: JmsProducerSession,
+                                      val epoch: Int) {
+
+  private val defaultDestination = jmsSession.jmsDestination
+
+  private val destinationCache = new SoftReferenceCache[Destination, jms.Destination]()
+
+  def send(elem: JmsMessage): Unit = {
+    val message: jms.Message = createMessage(elem)
+    populateMessageProperties(message, elem)
+
+    val (sendHeaders, headersBeforeSend: Set[JmsHeader]) = elem.headers.partition(_.usedDuringSend)
+    populateMessageHeader(message, headersBeforeSend)
+
+    val deliveryMode = sendHeaders
+      .collectFirst { case x: JmsDeliveryMode => x.deliveryMode }
+      .getOrElse(jmsProducer.getDeliveryMode)
+
+    val priority = sendHeaders
+      .collectFirst { case x: JmsPriority => x.priority }
+      .getOrElse(jmsProducer.getPriority)
+
+    val timeToLive = sendHeaders
+      .collectFirst { case x: JmsTimeToLive => x.timeInMillis }
+      .getOrElse(jmsProducer.getTimeToLive)
+
+    val destination = elem.destination match {
+      case Some(messageDestination) => lookup(messageDestination)
+      case None => defaultDestination
+    }
+    jmsProducer.send(destination, message, deliveryMode, priority, timeToLive)
+  }
+
+  private def lookup(dest: Destination) = destinationCache.lookup(dest, dest.create(jmsSession.session))
+
+  private[jms] def createMessage(element: JmsMessage): jms.Message =
+    element match {
+
+      case textMessage: JmsTextMessage => jmsSession.session.createTextMessage(textMessage.body)
+
+      case byteMessage: JmsByteMessage =>
+        val newMessage = jmsSession.session.createBytesMessage()
+        newMessage.writeBytes(byteMessage.bytes)
+        newMessage
+
+      case mapMessage: JmsMapMessage =>
+        val newMessage = jmsSession.session.createMapMessage()
+        populateMapMessage(newMessage, mapMessage)
+        newMessage
+
+      case objectMessage: JmsObjectMessage => jmsSession.session.createObjectMessage(objectMessage.serializable)
+
+    }
+
+  private[jms] def populateMessageProperties(message: javax.jms.Message, jmsMessage: JmsMessage): Unit =
+    jmsMessage.properties.foreach {
+      case (key, v) =>
+        v match {
+          case v: String => message.setStringProperty(key, v)
+          case v: Int => message.setIntProperty(key, v)
+          case v: Boolean => message.setBooleanProperty(key, v)
+          case v: Byte => message.setByteProperty(key, v)
+          case v: Short => message.setShortProperty(key, v)
+          case v: Long => message.setLongProperty(key, v)
+          case v: Double => message.setDoubleProperty(key, v)
+          case null => throw NullMessageProperty(key, jmsMessage)
+          case _ => throw UnsupportedMessagePropertyType(key, v, jmsMessage)
+        }
+    }
+
+  private def populateMapMessage(message: javax.jms.MapMessage, jmsMessage: JmsMapMessage): Unit =
+    jmsMessage.body.foreach {
+      case (key, v) =>
+        v match {
+          case v: String => message.setString(key, v)
+          case v: Int => message.setInt(key, v)
+          case v: Boolean => message.setBoolean(key, v)
+          case v: Byte => message.setByte(key, v)
+          case v: Short => message.setShort(key, v)
+          case v: Long => message.setLong(key, v)
+          case v: Double => message.setDouble(key, v)
+          case v: Array[Byte] => message.setBytes(key, v)
+          case null => throw NullMapMessageEntry(key, jmsMessage)
+          case _ => throw UnsupportedMapMessageEntryType(key, v, jmsMessage)
+        }
+    }
+
+  private def populateMessageHeader(message: javax.jms.Message, headers: Set[JmsHeader]): Unit =
+    headers.foreach {
+      case JmsType(jmsType) => message.setJMSType(jmsType)
+      case JmsReplyTo(destination) => message.setJMSReplyTo(destination.create(jmsSession.session))
+      case JmsCorrelationId(jmsCorrelationId) => message.setJMSCorrelationID(jmsCorrelationId)
+    }
+}
+
+/**
+ * Internal API.
+ */
+@InternalApi
+private[jms] object JmsMessageProducer {
+  def apply(jmsSession: JmsProducerSession, settings: JmsProducerSettings, epoch: Int): JmsMessageProducer = {
+    val producer = jmsSession.session.createProducer(null)
+    if (settings.timeToLive.nonEmpty) {
+      producer.setTimeToLive(settings.timeToLive.get.toMillis)
+    }
+    new JmsMessageProducer(producer, jmsSession, epoch)
+  }
+}
